Styled address field, implemented store logic
