+++ src/components/Editor/PostcardLayout.vue
@@ -90,14 +90,17 @@ export default Vue.extend({
   EventBus.$on('changeFont', (fontId: string) => {
     const textarea = document.querySelector('#changed-text') as HTMLElement;
     textarea.style.fontFamily = `${fontId}`;
-  })
+  }),
+  EventBus.$on('changeRecipient', () => { this.changeRecipient() })
 },
 methods: {
-    changeRecipient(recipient: Recipient){
+    changeRecipient(){
+      console.log("Die Methode wurde aufgerufen")
       const recipientName = document.getElementById('name') as HTMLInputElement;
       const recipientAddition = document.getElementById('addition') as HTMLInputElement;
       const recipientStreetAndNumber = document.getElementById('street-and-number') as HTMLInputElement;
       const recipientPostalcodeAndCity = document.getElementById('postalcode-and-city') as HTMLInputElement;
+      let recipient: Recipient;
       recipient = {
         name: recipientName.value,
         addition: recipientAddition.value,
@@ -106,13 +109,12 @@ methods: {
         country: ""
       }
       this.$store.dispatch("setRecipient", recipient)
-      console.log(this.$store)
     }
   },
   props: {
     ImageId: String,
   },
-  computed: {...mapState(["currentPicture", "currentBackgroundColor", "currentSticker", "currentTemplate"])
+  computed: {...mapState(["currentPicture", "currentBackgroundColor", "currentSticker", "currentTemplate", "currentRecipient"])
 }});
 </script>
 
@@ -192,7 +194,6 @@ methods: {
 .codierzone>p{
   margin: 10px 0 5px;
   text-align: center;
-   font-family: Montserrat, sans-serif !important;
 }
 
 .frankierzone {
@@ -214,7 +215,6 @@ methods: {
   width: 100%;
   padding-top: 25px;
   padding-left: 25px;
-  font-family: Dancing Script, serif !important;
 }
 
 .address-separator {
@@ -245,34 +245,31 @@ methods: {
   padding: 5px;
 }
 
+/* 
+.v-select__selection--comma {
+  font-size: 18px !important;
+  color: #707070 !important;
+  font-size: 14px !important;
+} */
+
 </style>
 <style scoped>
+
 
 .address-line {
   height: 35px;
   width: 200px;
   border-bottom: 1px solid #000;
-  font-family: Dancing Script, serif !important;
-  font-size: 18px;
+  font-size: 16px;
+  color: #000;
 }
 
 .address-line:focus {
  outline: 0 !important;
 }
 
-.v-select__selection--comma {
-  font-family: Dancing Script, serif !important;
-  font-size: 18px !important;
-  color: #707070 !important;
-}
-
-.v-input__control {
-  font-family: Dancing Script, serif !important;
-}
-
 .v-select__selections {
   padding: 0 !important;
-  font-family: Dancing Script, serif !important;
 }
 
 .v-select {
@@ -280,7 +277,6 @@ methods: {
   padding-top: 10 !important;
   padding-bottom: 10 !important;
   width: 200px;
-  font-family: Dancing Script, serif !important;
 }
 
 .v-text-field {
